Дебаг и доработка
- исправлено +2 ошибки постмана
- поиск событий теперь принимает LocalDateTime вместо String

+++ main-service/src/main/java/ru/practicum/ewm/service/request/EventRequestCount.java
@@ -5,5 +5,5 @@ public interface EventRequestCount {
 
     Long getEventId();
 
-    Integer getCount();
+    Long getCount();
 }
